third_party/blink/web_tests/crypto/random-values-expected.txt

CONSOLE WARNING: SharedArrayBuffer will require cross-origin isolation as of M91, around May 2021. See https://developer.chrome.com/blog/enabling-shared-array-buffer/ for more details.
Tests crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'crypto' in self is true
PASS 'getRandomValues' in self.crypto is true
PASS self.crypto.__proto__.hasOwnProperty('getRandomValues') is true
PASS matchingBytes < 100 is true
PASS crypto.getRandomValues(new Uint8Array(new SharedArrayBuffer(100))) threw exception TypeError: Failed to execute 'getRandomValues' on 'Crypto': The provided ArrayBufferView value must not be shared..
PASS successfullyParsed is true

TEST COMPLETE
